Split script to script_vm and script
