+++ src/script.c
@@ -12,12 +12,6 @@ extern u16 CalcCRC16WithTable(u8 *data, int length); // util
 #define RAM_SCRIPT_MAGIC 51
 
 enum {
-    SCRIPT_MODE_STOPPED,
-    SCRIPT_MODE_BYTECODE,
-    SCRIPT_MODE_NATIVE,
-};
-
-enum {
     CONTEXT_RUNNING,
     CONTEXT_WAITING,
     CONTEXT_SHUTDOWN,
@@ -40,159 +34,6 @@ static u8 sMsgIsSignpost;
 
 extern ScrCmdFunc gScriptCmdTable[];
 extern ScrCmdFunc gScriptCmdTableEnd[];
-extern void *gNullScriptPtr;
-
-void InitScriptContext(struct ScriptContext *ctx, void *cmdTable, void *cmdTableEnd)
-{
-    s32 i;
-
-    ctx->mode = SCRIPT_MODE_STOPPED;
-    ctx->scriptPtr = NULL;
-    ctx->stackDepth = 0;
-    ctx->nativePtr = NULL;
-    ctx->cmdTable = cmdTable;
-    ctx->cmdTableEnd = cmdTableEnd;
-
-    for (i = 0; i < (int)ARRAY_COUNT(ctx->data); i++)
-        ctx->data[i] = 0;
-
-    for (i = 0; i < (int)ARRAY_COUNT(ctx->stack); i++)
-        ctx->stack[i] = 0;
-}
-
-u8 SetupBytecodeScript(struct ScriptContext *ctx, const u8 *ptr)
-{
-    ctx->scriptPtr = ptr;
-    ctx->mode = SCRIPT_MODE_BYTECODE;
-    return 1;
-}
-
-void SetupNativeScript(struct ScriptContext *ctx, bool8 (*ptr)(void))
-{
-    ctx->mode = SCRIPT_MODE_NATIVE;
-    ctx->nativePtr = ptr;
-}
-
-void StopScript(struct ScriptContext *ctx)
-{
-    ctx->mode = SCRIPT_MODE_STOPPED;
-    ctx->scriptPtr = NULL;
-}
-
-bool8 RunScriptCommand(struct ScriptContext *ctx)
-{
-    // FRLG disabled this check, where-as it is present
-    // in Ruby/Sapphire and Emerald. Why did the programmers
-    // bother to remove a redundant check when it still
-    // exists in Emerald?
-    //if (ctx->mode == SCRIPT_MODE_STOPPED)
-    //    return FALSE;
-
-    switch (ctx->mode)
-    {
-    case SCRIPT_MODE_STOPPED:
-        return FALSE;
-    case SCRIPT_MODE_NATIVE:
-        // Try to call a function in C
-        // Continue to bytecode if no function or it returns TRUE
-        if (ctx->nativePtr)
-        {
-            if (ctx->nativePtr() == TRUE)
-                ctx->mode = SCRIPT_MODE_BYTECODE;
-            return TRUE;
-        }
-        ctx->mode = SCRIPT_MODE_BYTECODE;
-        // fallthrough
-    case SCRIPT_MODE_BYTECODE:
-        while (1)
-        {
-            u8 cmdCode;
-            ScrCmdFunc *cmdFunc;
-
-            if (ctx->scriptPtr == NULL)
-            {
-                ctx->mode = SCRIPT_MODE_STOPPED;
-                return FALSE;
-            }
-
-            if (ctx->scriptPtr == gNullScriptPtr)
-            {
-                while (1)
-                    asm("svc 2"); // HALT
-            }
-
-            cmdCode = *(ctx->scriptPtr);
-            ctx->scriptPtr++;
-            cmdFunc = &ctx->cmdTable[cmdCode];
-
-            if (cmdFunc >= ctx->cmdTableEnd)
-            {
-                ctx->mode = SCRIPT_MODE_STOPPED;
-                return FALSE;
-            }
-
-            if ((*cmdFunc)(ctx) == TRUE)
-                return TRUE;
-        }
-    }
-
-    return TRUE;
-}
-
-static u8 ScriptPush(struct ScriptContext *ctx, const u8 *ptr)
-{
-    if (ctx->stackDepth + 1 >= (int)ARRAY_COUNT(ctx->stack))
-    {
-        return 1;
-    }
-    else
-    {
-        ctx->stack[ctx->stackDepth] = ptr;
-        ctx->stackDepth++;
-        return 0;
-    }
-}
-
-static const u8 *ScriptPop(struct ScriptContext *ctx)
-{
-    if (ctx->stackDepth == 0)
-        return NULL;
-
-    ctx->stackDepth--;
-    return ctx->stack[ctx->stackDepth];
-}
-
-void ScriptJump(struct ScriptContext *ctx, const u8 *ptr)
-{
-    ctx->scriptPtr = ptr;
-}
-
-void ScriptCall(struct ScriptContext *ctx, const u8 *ptr)
-{
-    ScriptPush(ctx, ctx->scriptPtr);
-    ctx->scriptPtr = ptr;
-}
-
-void ScriptReturn(struct ScriptContext *ctx)
-{
-    ctx->scriptPtr = ScriptPop(ctx);
-}
-
-u16 ScriptReadHalfword(struct ScriptContext *ctx)
-{
-    u16 value = *(ctx->scriptPtr++);
-    value |= *(ctx->scriptPtr++) << 8;
-    return value;
-}
-
-u32 ScriptReadWord(struct ScriptContext *ctx)
-{
-    u32 value0 = *(ctx->scriptPtr++);
-    u32 value1 = *(ctx->scriptPtr++);
-    u32 value2 = *(ctx->scriptPtr++);
-    u32 value3 = *(ctx->scriptPtr++);
-    return (((((value3 << 8) + value2) << 8) + value1) << 8) + value0;
-}
 
 void LockPlayerFieldControls(void)
 {
