+++ sym_ewram_rev10.txt
@@ -21,6 +21,7 @@
 	.include "src/field_player_avatar.o"
 	.include "src/event_object_movement.o"
 	.include "src/field_message_box.o"
+	.include "src/script_vm.o"
 	.include "src/script.o"
 	.include "src/scrcmd.o"
 	.include "src/event_data.o"
